Refactor icicle charts into a single generic component
- Created ObjectsIcicleChart.tsx as a generic reusable icicle chart component
- Moved data wrangling logic from chart components to AccountTab, IncomeTab, and ExpenseTab
- Deleted AccountIcicleChart.tsx, IncomeIcicleChart.tsx, and ExpenseIcicleChart.tsx
- Follows same pattern as CashflowSankey refactoring
- Reduces code duplication and improves maintainability
