+++ src/tabs/Current/ExpenseTab.tsx
@@ -1,12 +1,18 @@
-import { useState, useContext } from 'react';
+import { useState, useContext, useMemo } from 'react';
 import { ExpenseContext } from '../../components/Objects/Expense/ExpenseContext';
-import { 
-    BaseExpense
+import {
+    BaseExpense,
+    AnyExpense,
+    LoanExpense,
+    CLASS_TO_CATEGORY,
+    CATEGORY_PALETTES,
+    EXPENSE_CATEGORIES,
+    isExpenseActiveInCurrentMonth
 } from '../../components/Objects/Expense/models';
 import ExpenseCard from '../../components/Objects/Expense/ExpenseCard';
 import AddExpenseModal from '../../components/Objects/Expense/AddExpenseModal';
 import { DragDropContext, Droppable, Draggable, DropResult } from '@hello-pangea/dnd';
-import ExpenseIcicleChart from '../../components/Objects/Expense/ExpenseIcicleChart';
+import { ObjectsIcicleChart, tailwindToCssVar, getDistributedColors } from '../../components/Charts/ObjectsIcicleChart';
 
 const ExpenseList = ({ type }: { type: any }) => {
   const { expenses, dispatch } = useContext(ExpenseContext);
@@ -74,6 +80,51 @@ const TabsContent = () => {
     const { expenses } = useContext(ExpenseContext);
     const [isModalOpen, setIsModalOpen] = useState(false);
 
+    // Data wrangling for icicle chart
+    const hierarchicalData = useMemo(() => {
+        const grouped: Record<string, AnyExpense[]> = {};
+
+        // 1. Group expenses (only active ones)
+        expenses
+            .filter(isExpenseActiveInCurrentMonth)
+            .forEach((exp) => {
+                const category = CLASS_TO_CATEGORY[exp.constructor.name] || 'Other';
+                if (!grouped[category]) grouped[category] = [];
+                grouped[category].push(exp);
+            });
+
+        // 2. Build Children with Colors
+        const categoryChildren = EXPENSE_CATEGORIES.map((category) => {
+            const expensesInCategory = grouped[category] || [];
+            if (expensesInCategory.length === 0) return null;
+
+            // Get gradient colors for this specific group of expenses
+            const palette = CATEGORY_PALETTES[category];
+            const expenseColors = getDistributedColors(palette, expensesInCategory.length);
+            // Pick a representative color for the Category header (approx middle of palette)
+            const categoryColor = palette[50] || palette[Math.floor(palette.length/2)];
+
+            return {
+                id: category,
+                color: tailwindToCssVar(categoryColor), // Parent Color
+                children: expensesInCategory.map((exp, i) => ({
+                    id: exp.name,
+                    value: exp.getMonthlyAmount(),
+                    color: tailwindToCssVar(expenseColors[i]), // Child Gradient Color
+                    // Metadata
+                    originalAmount: exp instanceof LoanExpense ? exp.payment : exp.amount,
+                    frequency: exp.frequency
+                }))
+            };
+        }).filter(Boolean); // Remove empty categories
+
+        return {
+            id: "Total Expenses",
+            color: "#ef4444", // Root node color
+            children: categoryChildren
+        };
+    }, [expenses]);
+
     return (
         <div className="w-full min-h-full flex bg-gray-950 justify-center pt-6">
             <div className="w-full px-8 max-w-screen-2xl">
@@ -84,7 +135,9 @@ const TabsContent = () => {
                     </h2>
                     
                     {expenses.length > 0 && (
-                        <ExpenseIcicleChart expenseList= {expenses}
+                        <ObjectsIcicleChart
+                            data={hierarchicalData}
+                            valueFormat=">-$0,.0f"
                         />
                     )}
                 </div>
